feat(ibd): fixing pr
